Removed nix-gaming flake and swapped /media to btrfs
proton-ge-bin was added to unstable and works with extraCompatPackages, so this commit swaps to that implementation and nix-gaming was removed. The /Media partition was converted to btrfs to take advantage of zstd compression saving 140gb of disk space. Other smaller changes are fixing deprecations in option names and temporarily removing super-scicer as it is not building due to an upstream regression.